perf(ui): Agent 执行面板性能优化与 token 统计修正
- 抽离 ElapsedTimer：每秒计时 tick 不再触发整条 trace 重渲
- ModelPlanGroup 改用父组件按 step 预分组的事件切片，消除每个分组各自遍历整条 trace 的 O(n²)
- AgentPanel 派生值 useMemo 化；单条 trace 抽为 React.memo 的 TraceItem，避免重复 JSON.stringify
- 统一 computeTraceMetrics：按 step 取 model_plan 决策 usage（无则回退 action），修复 model_plan 与 action 重复计数（单模型约 2×）及桌面/移动端口径不一致
- 新增 useIsMobile（响应 resize）替换一次性 innerWidth 判定；流式滚动改 behavior:auto 防抖动
