+++ client/src/components/AgentPane.jsx
@@ -1,24 +1,5 @@
 import { useT } from '../i18n/I18nProvider.jsx';
-
-function getAgentStepMetrics(trace) {
-  const lastStep = trace.reduce((max, event) => (event.step != null ? Math.max(max, event.step) : max), 0);
-  const totalTokens = trace.reduce((sum, event) => {
-    if (event.type === 'step' && event.stage === 'action' && event.usage) {
-      return sum + event.usage.prompt_tokens + event.usage.completion_tokens;
-    }
-    return sum;
-  }, 0);
-  const doneEvent = [...trace].reverse().find(event => event.type === 'done');
-  return {
-    lastStep,
-    totalTokens,
-    stepCount: doneEvent?.meta?.step_count || lastStep,
-  };
-}
-
-function formatTokenCount(value) {
-  return value > 999 ? `${(value / 1000).toFixed(1)}k` : value;
-}
+import { computeTraceMetrics, formatTokenCount } from './agent/trace-metrics.js';
 
 export function AgentPane({
   agentMobileTab,
@@ -30,7 +11,7 @@ export function AgentPane({
   resizeDivider,
 }) {
   const t = useT();
-  const metrics = getAgentStepMetrics(agentTrace);
+  const metrics = computeTraceMetrics(agentTrace);
 
   return (
     <>
